+++ release-notes.md
@@ -1,4 +1,11 @@
 # Propus 16.7.0  #
+## Adding privileges field on widget object ##
+Issue Type: Task
+Issue ID : PLAT-10953
+
+#### Deployment Scripts ####
+php /opt/kaltura/app/deployment/updates/scripts/add_permissions/2020_07_26_add_permission_widget_object_privileges_field.php
+
 ## Add more user details name to Q&A response profile  ##
 - Issue Type: Feature
 - Issue ID: WEBC-1925